dev 1、修改前置插件的下载地址(提示) 2、将tipsServer.java tipsPlayer.java tipsText.java这个三个类的内容统一放在BaseVariables.java中
